+++ windows-azure-storage-settings.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Shows various settings for Windows Azure Storage Plugin
+ * Shows various settings for Microsoft Azure Storage Plugin
  *
  * Version: 3.0.1
  *
@@ -33,7 +33,7 @@
  * PHP Version 5
  *
  * @category  WordPress_Plugin
- * @package   Windows_Azure_Storage_For_WordPress
+ * @package   Microsoft_Azure_Storage_For_WordPress
  * @author    Microsoft Open Technologies, Inc. <msopentech@microsoft.com>
  * @copyright Microsoft Open Technologies, Inc.
  * @license   New BSD license, (http://www.opensource.org/licenses/bsd-license.php)
@@ -41,7 +41,7 @@
  */
 
 /**
- * Preamble text on Windows Azure Storage plugin settings page.
+ * Preamble text on Microsoft Azure Storage plugin settings page.
  *
  * @since 4.0.0
  *
@@ -51,15 +51,16 @@ function windows_azure_storage_plugin_settings_preamble() {
 	?>
 	<div class="wrap">
 		<h2>
-			<img src="<?php echo esc_url( MSFT_AZURE_PLUGIN_URL . 'images/WindowsAzure.jpg' ); ?>" width="32" height="32"/><?php esc_html_e( 'Windows Azure Storage for WordPress', 'windows-azure-storage' ); ?>
+			<img src="<?php echo esc_url( MSFT_AZURE_PLUGIN_URL . 'images/azure-icon.png' ); ?>" alt="<?php esc_attr_e( 'Microsoft Azure', 'windows-azure-storage' ); ?>" style="width:32px">
+			<?php esc_html_e( 'Microsoft Azure Storage for WordPress', 'windows-azure-storage' ); ?>
 		</h2>
 
 		<?php esc_html_e(
-			'This WordPress plugin allows you to use Windows Azure Storage Service to host your media for your WordPress powered blog. Windows Azure provides storage in the cloud with authenticated access and triple replication to help keep your data safe. Applications work with data using REST conventions and standard HTTP operations to identify and expose data using URIs. This plugin allows you to easily upload, retrieve, and link to files stored on Windows Azure Storage service from within WordPress.',
+			'This WordPress plugin allows you to use Microsoft Azure Storage Service to host your media for your WordPress powered blog. Microsoft Azure provides storage in the cloud with authenticated access and triple replication to help keep your data safe. Applications work with data using REST conventions and standard HTTP operations to identify and expose data using URIs. This plugin allows you to easily upload, retrieve, and link to files stored on Microsoft Azure Storage service from within WordPress.',
 			'windows-azure-storage'
 		); ?>
 		<br/><br/>
-		<?php echo __( 'For more details on Windows Azure Storage Services, please visit the <a href="http://www.microsoft.com/azure/windowsazure.mspx">Windows Azure Platform web-site</a>.', 'windows-azure-storage' ); ?>
+		<?php echo __( 'For more details on Microsoft Azure Storage Services, please visit the <a href="http://www.microsoft.com/azure/windowsazure.mspx">Microsoft Azure Platform web-site</a>.', 'windows-azure-storage' ); ?>
 		<br/>
 
 		<b><?php esc_html_e( 'Plugin Web Site:', 'windows-azure-storage' ); ?></b>
@@ -93,7 +94,7 @@ function windows_azure_storage_plugin_options_page() {
 }
 
 /**
- * Register custom settings for Windows Azure Storage Plugin.
+ * Register custom settings for Microsoft Azure Storage Plugin.
  *
  * @return void
  */
@@ -112,7 +113,7 @@ function windows_azure_storage_plugin_register_settings() {
 	 */
 	add_settings_section(
 		'windows-azure-storage-settings',
-		__( 'Windows Azure Storage Settings', 'windows-azure-storage' ),
+		__( 'Microsoft Azure Storage Settings', 'windows-azure-storage' ),
 		'windows_azure_storage_plugin_settings_section',
 		'windows-azure-storage-plugin-options'
 	);
@@ -161,7 +162,7 @@ function windows_azure_storage_plugin_register_settings() {
 	 */
 	add_settings_field(
 		'azure_storage_handle_uploads',
-		__( 'Use Windows Azure Storage for default upload', 'windows-azure-storage' ),
+		__( 'Use Microsoft Azure Storage for default upload', 'windows-azure-storage' ),
 		'windows_azure_storage_setting_handle_uploads',
 		'windows-azure-storage-plugin-options',
 		'windows-azure-storage-settings'
@@ -207,7 +208,7 @@ function windows_azure_storage_plugin_register_settings() {
  */
 function windows_azure_storage_plugin_settings_section() {
 	?>
-	<p><?php echo __( 'If you do not have Windows Azure Storage Account, please <a href="http://go.microsoft.com/fwlink/?LinkID=129453">register </a>for Windows Azure Services.', 'windows-azure-storage' ); ?></p>
+	<p><?php echo __( 'If you do not have Microsoft Azure Storage Account, please <a href="http://go.microsoft.com/fwlink/?LinkID=129453">register </a>for Microsoft Azure Services.', 'windows-azure-storage' ); ?></p>
 	<?php
 }
 
@@ -222,7 +223,7 @@ function windows_azure_storage_plugin_settings_section() {
 function windows_azure_storage_setting_account_name() {
 	$storage_account_name = Windows_Azure_Helper::get_account_name();
 	?>
-	<input type="text" name="azure_storage_account_name" class="regular-text" title="<?php esc_attr_e( 'Windows Azure Storage Account Name', 'windows-azure-storage' ); ?>" value="<?php echo esc_attr( $storage_account_name ); ?>"/>
+	<input type="text" name="azure_storage_account_name" class="regular-text" title="<?php esc_attr_e( 'Microsoft Azure Storage Account Name', 'windows-azure-storage' ); ?>" value="<?php echo esc_attr( $storage_account_name ); ?>"/>
 	<?php
 }
 
@@ -236,7 +237,7 @@ function windows_azure_storage_setting_account_name() {
 function windows_azure_storage_setting_account_key() {
 	$storage_account_key = Windows_Azure_Helper::get_account_key();
 	?>
-	<input type="text" name="azure_storage_account_primary_access_key" class="large-text" title="<?php esc_attr_e( 'Windows Azure Storage Account Primary Access Key', 'windows-azure-storage' ); ?>" value="<?php echo esc_attr( $storage_account_key ); ?>"/>
+	<input type="text" name="azure_storage_account_primary_access_key" class="large-text" title="<?php esc_attr_e( 'Microsoft Azure Storage Account Primary Access Key', 'windows-azure-storage' ); ?>" value="<?php echo esc_attr( $storage_account_key ); ?>"/>
 	<?php
 }
 
@@ -305,7 +306,7 @@ function windows_azure_storage_setting_storage_container() {
 function windows_azure_storage_setting_cname() {
 	$cname = Windows_Azure_Helper::get_cname();
 	?>
-	<input type="url" name="cname" class="regular-text" title="<?php esc_attr_e( 'Use CNAME instead of Windows Azure Blob URL', 'windows-azure-storage' ); ?>" value="<?php echo esc_attr( $cname ); ?>"/>
+	<input type="url" name="cname" class="regular-text" title="<?php esc_attr_e( 'Use CNAME instead of Microsoft Azure Blob URL', 'windows-azure-storage' ); ?>" value="<?php echo esc_attr( $cname ); ?>"/>
 	<p class="field-description">
 		<?php _e( 'Note: Use this option if you would like to display image URLs belonging to your domain like <code>http://mydomain.com/</code> instead of <code>http://your-account-name.blob.core.windows.net/</code>.', 'windows-azure-storage' ); ?>
 	</p>
@@ -324,9 +325,9 @@ function windows_azure_storage_setting_cname() {
  */
 function windows_azure_storage_setting_handle_uploads() {
 	?>
-	<input type="checkbox" name="azure_storage_use_for_default_upload" title="<?php esc_attr_e( 'Use Windows Azure Storage for default upload', 'windows-azure-storage' ) ?>" value="1" id="azure_storage_use_for_default_upload" <?php checked( (bool) get_option( 'azure_storage_use_for_default_upload' ) ); ?> />
+	<input type="checkbox" name="azure_storage_use_for_default_upload" title="<?php esc_attr_e( 'Use Microsoft Azure Storage for default upload', 'windows-azure-storage' ) ?>" value="1" id="azure_storage_use_for_default_upload" <?php checked( (bool) get_option( 'azure_storage_use_for_default_upload' ) ); ?> />
 	<label for="azure_storage_use_for_default_upload">
-		<?php esc_html_e( 'Use Windows Azure Storage when uploading via WordPress\' upload tab.', 'windows-azure-storage' ); ?>
+		<?php esc_html_e( 'Use Microsoft Azure Storage when uploading via WordPress\' upload tab.', 'windows-azure-storage' ); ?>
 	</label>
 	<br/>
 	<small><?php esc_html_e( 'Note: Uncheck this to revert back to using your own web host for storage at anytime.', 'windows-azure-storage' ); ?></small>
@@ -383,7 +384,7 @@ function windows_azure_cache_control() {
 	
 	?><input type="number" name="azure_cache_control" class="regular-text" value="<?php echo esc_attr( $cache_control ); ?>">
 	<p class="field-description">
-		<?php esc_html_e( 'Setting Cache-Control on publicly accessible Windows Azure Blobs can help reduce bandwidth by preventing consumers from having to continuously download resources. Specify a relative amount of time in seconds to cache data after it was received.', 'windows-azure-storage' ); ?>
+		<?php esc_html_e( 'Setting Cache-Control on publicly accessible Microsoft Azure Blobs can help reduce bandwidth by preventing consumers from having to continuously download resources. Specify a relative amount of time in seconds to cache data after it was received.', 'windows-azure-storage' ); ?>
 	</p><?php	
 }